Update writeClip.R
Switched the `stop` to a `warning` as if this function is used in a Vignette or example it will throw an error on a Linux box without xclip installed.  The warning laxes this a bit.  I can't see of a reason to use the `stop` over the `warning` but there may be and I haven't thought of it.